Extracted unregistered repo dropdown to separate module
