+++ src/common/components/repository-row/dropdowns/index.js
@@ -1,9 +1,11 @@
 import NameMismatchDropdown from './name-mismatch-dropdown';
 import RemoveRepoDropdown from './remove-repo-dropdown';
 import UnconfiguredDropdown from './unconfigured-dropdown';
+import UnregisteredDropdown from './unregistered-dropdown';
 
 export {
   NameMismatchDropdown,
   RemoveRepoDropdown,
-  UnconfiguredDropdown
+  UnconfiguredDropdown,
+  UnregisteredDropdown
 };
